+++ CHANGELOG.md
@@ -14,8 +14,8 @@
 
 * Forked `x/gov` module from Cosmos SDK v0.46.16 for custom modifications
   ([#4](https://github.com/atomone-hub/govgen/pull/4)).
-* Removed ability for validators to vote on proposals
-  ([#10](https://github.com/atomone-hub/govgen/pull/10)).
+* Removed ability for validators to vote on proposals with delegations, they can only use their own stake
+  ([#20](https://github.com/atomone-hub/govgen/pull/20)).
 * Remove community spend proposal
   ([#13](https://github.com/atomone-hub/govgen/pull/13)).
 * Adapt voting period to proposal type.